feat: adds firebase, fix android build
