+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.flashcards",
+    package = "app.myflashcards.myflashcards",
 )
 
 android_resource(
     name = "res",
-    package = "com.flashcards",
+    package = "app.myflashcards.myflashcards",
     res = "src/main/res",
 )
