+++ android/app/src/main/java/app/myflashcards/myflashcards/MainActivity.java
@@ -1,4 +1,4 @@
-package com.flashcards;
+package app.myflashcards.myflashcards;
 
 import com.facebook.react.ReactActivity;
