+++ android/app/src/main/java/app/myflashcards/myflashcards/MainApplication.java
@@ -1,14 +1,13 @@
-package com.flashcards;
+package app.myflashcards.myflashcards;
 
 import android.app.Application;
 import android.content.Context;
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
-import com.flashcards.modules.FlashCardsAppVersionPackage;
-import com.flashcards.modules.FlashCardsRateAppPackage;
-import com.ijzerenhein.sharedelement.RNSharedElementPackage;
-import com.ijzerenhein.sharedelement.RNSharedElementPackage;
-import com.th3rdwave.safeareacontext.SafeAreaContextPackage;
+
+import app.myflashcards.myflashcards.BuildConfig;
+import app.myflashcards.myflashcards.modules.FlashCardsAppVersionPackage;
+import app.myflashcards.myflashcards.modules.FlashCardsRateAppPackage;
 import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
@@ -32,7 +31,6 @@ public class MainApplication extends Application implements ReactApplication {
           // Packages that cannot be autolinked yet can be added manually here, for example:
           // packages.add(new MyReactNativePackage());
           packages.add(new FlashCardsRateAppPackage());
-          packages.add(new SafeAreaContextPackage());
           packages.add(new FlashCardsAppVersionPackage());
           return packages;
         }
@@ -70,7 +68,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.flashcards.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("app.myflashcards.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
