+++ android/app/src/main/java/app/myflashcards/myflashcards/modules/FlashCardsAppVersionModule.java
@@ -1,4 +1,4 @@
-package com.flashcards.modules;
+package app.myflashcards.myflashcards.modules;
 
 import com.facebook.react.BuildConfig;
 import com.facebook.react.bridge.Arguments;
