+++ android/app/src/main/java/app/myflashcards/myflashcards/modules/FlashCardsRateAppPackage.java
@@ -1,23 +1,24 @@
-package com.flashcards.modules;
+package app.myflashcards.myflashcards.modules;
 
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.ViewManager;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.ArrayList;
 
-public class FlashCardsAppVersionPackage implements ReactPackage {
+public class FlashCardsRateAppPackage implements ReactPackage {
     @Override
     public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
         return Collections.emptyList();
     }
-
-    @Override public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
+    @Override
+    public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
         List<NativeModule> modules = new ArrayList<>();
-        modules.add(new FlashCardsAppVersionModule(reactContext));
+
+        modules.add(new FlashCardsRateAppModule(reactContext));
         return modules;
     }
 }
