+++ android/build.gradle
@@ -13,6 +13,7 @@ buildscript {
     }
     dependencies {
         classpath("com.android.tools.build:gradle:3.5.3")
+        classpath("com.google.gms:google-services:4.3.4")
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
