+++ ios/GoogleService-Info.plist
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>408639721386-3g07bmsb7ae8vlji5r0tv869kejib2pj.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.408639721386-3g07bmsb7ae8vlji5r0tv869kejib2pj</string>
+	<key>API_KEY</key>
+	<string>AIzaSyC4jR1L_g1PjriXhS-pCerwET6TrlTygVY</string>
+	<key>GCM_SENDER_ID</key>
+	<string>408639721386</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>app.myflashcards.myflashcards</string>
+	<key>PROJECT_ID</key>
+	<string>myflashcards-c32f1</string>
+	<key>STORAGE_BUCKET</key>
+	<string>myflashcards-c32f1.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:408639721386:ios:b27066add798aa3e7ca5d3</string>
+</dict>
+</plist>
