+++ package.json
@@ -10,8 +10,9 @@
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
   "dependencies": {
-    "@react-native-community/async-storage": "^1.12.1",
+    "@react-native-async-storage/async-storage": "^1.14.1",
     "@react-native-community/masked-view": "^0.1.10",
+    "@react-native-firebase/app": "^11.1.1",
     "@react-native-html/parser": "^0.1.0",
     "@react-navigation/bottom-tabs": "^5.11.2",
     "@react-navigation/drawer": "^5.11.4",
@@ -35,7 +36,7 @@
     "react-native-image-picker": "^3.1.4",
     "react-native-pell-rich-editor": "^1.6.3",
     "react-native-reanimated": "^1.13.2",
-    "react-native-safe-area-context": "^3.1.9",
+    "react-native-safe-area-context": "^3.2.0",
     "react-native-screens": "^2.15.0",
     "react-native-share": "^5.1.1",
     "react-native-shared-element": "^0.7.0",
