+++ src/redux/store.ts
@@ -1,7 +1,7 @@
 import { applyMiddleware, combineReducers, createStore, Middleware } from 'redux';
 import { createLogger } from 'redux-logger';
 import { persistStore, persistReducer } from 'redux-persist';
-import AsyncStorage from '@react-native-community/async-storage';
+import AsyncStorage from '@react-native-async-storage/async-storage';
 import autoMergeLevel2 from 'redux-persist/lib/stateReconciler/autoMergeLevel2';
 import createSagaMiddleware from 'redux-saga';
 import decks, { DecksState } from './decks/reducer';
